+++ demo/src/pages/mod.rs
@@ -3,6 +3,7 @@
 mod component_page;
 mod getting_started;
 mod home;
+mod node_graph;
 mod not_found;
 mod theming_custom;
 mod theming_overview;
@@ -11,6 +12,7 @@ mod theming_presets;
 pub use component_page::*;
 pub use getting_started::*;
 pub use home::*;
+pub use node_graph::*;
 pub use not_found::*;
 pub use theming_custom::*;
 pub use theming_overview::*;